requirements/testing.txt
-r base.txt

black==19.3b0
pylint==2.3.1
pytest==5.0.1
pytest-cov==2.7.1
pytest-aiohttp==0.3.0
tox==3.13.2
mypy==0.720
asynctest==0.13.0
mock==3.0.5
